Refactor audio message event tests by moving mock implementations to a new helper file. This improves test organization and maintains clarity in the audio_message_event_test.dart file.
